+++ Sources/svglib/line.cpp
@@ -18,6 +18,6 @@ void Line::set_stroke(Stroke s) {
 string Line::generate(float x1, float y1, float x2, float y2) {
     string stro;
     stro.assign(stroke.stroke);
-    string s = "<line stroke=\""+ stro + "\" stroke-width=\"" + to_string((int) stroke.stroke_width) + "\" x1=\"" + to_string((int) x1) + "\" x2=\"" + to_string((int) x2) + "\" y1=\"" + to_string((int) y1) +"\" y2=\"" + to_string((int) y2) +"\"/>";
+    string s = "<line stroke=\""+ stro + "\" stroke-width=\"" + to_string((int) stroke.stroke_width) + "\" x1=\"" + to_string(x1) + "\" x2=\"" + to_string(x2) + "\" y1=\"" + to_string(y1) +"\" y2=\"" + to_string(y2) +"\"/>";
     return s;
 }
